* Store handle in 64 bit integer instead of in 32 bit  * Changed some _WIN32_WCE to _MSC_VER where they are    always needed when compiling with visual c (not only    for Windows Mobile).  * Fixed some warnings for MSVC   - Deprecated functions: sprintf, _ftime
git-svn-id: svn://svn.minisip.org/minisip/trunk@3218 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/source/TextUI.cxx
@@ -165,7 +165,7 @@ void TextUI::answerQuestion(string answer){
 	MRef<QuestionDialog *> d = questionDialogs.front();
 	d->answers.push_back(answer);
 	d->nAnswered++;
-	if (d->nAnswered >= d->questions.size()){
+	if (d->nAnswered >= (int)d->questions.size()){
 		//all questions now answered
 		guiExecute(d);
 		isAskingDialogQuestion=false;
